/*超级棒棒糖*/

+++ src/main/java/com/controller/ListenController.java
@@ -1,5 +1,6 @@
 package com.controller;
 
+import com.alibaba.excel.EasyExcel;
 import com.github.pagehelper.PageInfo;
 import com.pojo.Listen;
 import com.service.ListenService;
@@ -9,8 +10,14 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.net.URLEncoder;
+import java.text.SimpleDateFormat;
 import java.util.Arrays;
+import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 
 @Controller
 public class ListenController {
@@ -18,9 +25,9 @@ public class ListenController {
     @Autowired
     private ListenService listenService;
 
-    /*
-    后台
-    */
+    /**
+     * 后台功能
+     */
     @GetMapping("/listenList")
     public String listenList(Listen listen, Model model) {
         model.addAttribute("info", listen);
@@ -57,8 +64,8 @@ public class ListenController {
                     Listen listen = new Listen();
                     listen.setSupNum(Integer.valueOf(supNum));
                     listen.setCourseNum(Integer.valueOf(course));
-                    listen.setVerifyState("2");
-                    listen.setListenState("1");
+                    listen.setVerifyState("0");
+                    listen.setListenState("0");
                     listenService.addListen(listen);
                 }
             } else if (courseList.length == 1 && supNumList.length > 0) {
@@ -66,8 +73,8 @@ public class ListenController {
                     Listen listen = new Listen();
                     listen.setSupNum(Integer.valueOf(supervisor));
                     listen.setCourseNum(Integer.valueOf(courseNum));
-                    listen.setVerifyState("2");
-                    listen.setListenState("1");
+                    listen.setVerifyState("0");
+                    listen.setListenState("0");
                     listenService.addListen(listen);
                 }
             } else {
@@ -111,10 +118,14 @@ public class ListenController {
         List<Listen> listenList = listenService.queryListenAllWithSame(sList, cList);
 
         if (listenList.size() != 0) {
-            if (supNumList.length == 1 && courseList.length > 0) {
-                listenService.deleteListenBySupNumAndCourseNum(Integer.valueOf(supNum), cList);
+            if (!Objects.equals(listenList.get(0).getVerifyState(), "1")) {
+                if (supNumList.length == 1 && courseList.length > 0) {
+                    listenService.deleteListenBySupNumAndCourseNum(Integer.valueOf(supNum), cList);
+                } else {
+                    return DataInfo.fail("课程取消报名失败，请联系督导秘书");
+                }
             } else {
-                return DataInfo.fail("课程取消报名失败，请联系督导秘书");
+                return DataInfo.fail("该课程已通过审核");
             }
         } else {
             return DataInfo.fail("未报名课程");
@@ -136,9 +147,9 @@ public class ListenController {
         return DataInfo.ok();
     }
 
-    /*
-    前台
-    */
+    /**
+     * 前台功能
+     */
     @GetMapping("/listenView")
     public String listenView() {
         return "listen/listenView";
@@ -176,5 +187,24 @@ public class ListenController {
     @GetMapping("/listenEvaluation")
     public String listenEvaluation() {
         return "listen/listenEvaluation";
+    }
+
+    @GetMapping("/downloadByListen")
+    public void downloadByListen(Integer[] listenNum, HttpServletResponse response) throws IOException {
+        if (listenNum.length > 0) {
+            //查询要写入的集合对象
+            List<Listen> listenList = listenService.queryListenByListenNums(listenNum);
+            System.out.println(listenList);
+            //写入Excel
+            System.out.println("系统时间: " + System.currentTimeMillis());
+            //获取当前时间
+            String time = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss").format(new Date());
+            //以下载的方式写入Excel
+            String fileName = URLEncoder.encode("已选课程", "UTF-8").replaceAll("\\+", "%20");
+            response.setHeader("Content-disposition",
+                    "attachment;filename*=utf-8''" + fileName + "-" + time + ".xlsx");
+            //注意写入的对象是：outputStream
+            EasyExcel.write(response.getOutputStream(), Listen.class).sheet("听课列表").doWrite(listenList);
+        }
     }
 }
